feat : map multipartfile test

+++ src/main/java/com/team6/sole/domain/home/entity/Course.java
@@ -37,7 +37,7 @@ public class Course extends BaseTimeEntity {
 
     private int duration;
 
-    private int distance;
+    private double distance;
 
     @ElementCollection(fetch = FetchType.LAZY) @Enumerated(EnumType.STRING)
     Set<PlaceCategory> placeCategories;
